Minor changes in post-processing. Added propper passthrough shader.

+++ src/null-render.h
@@ -16,7 +16,7 @@
 #include "null-render/backend/backend.h"
 
 namespace null::render {
-	void initialize(const vec2_t<float>& viewport);
+	void initialize(const vec2_t<int>& viewport);
 
 	void begin_frame();
 	void end_frame();
